+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppsworkload/integration/staff/GetStaffById.kt
@@ -2,12 +2,13 @@ package uk.gov.justice.digital.hmpps.hmppsworkload.integration.staff
 
 import org.junit.jupiter.api.Test
 import uk.gov.justice.digital.hmpps.hmppsworkload.integration.IntegrationTestBase
+import java.math.BigInteger
 
 class GetStaffById : IntegrationTestBase() {
 
   @Test
   fun `get staff by ID`() {
-    val staffId = 123456789L
+    val staffId = BigInteger.valueOf(123456789L)
     staffIdResponse(staffId, "OM1", "T1")
     webTestClient.get()
       .uri("/staff/$staffId")
